chore: 修复大量文档和代码细节问题
1. 修复多处中文注释的乱码字符
2. 修正测试代码中未使用变量的警告处理
3. 调整CMake依赖检测逻辑，使用统一全局缓存变量
4. 优化日志打印格式和中文表述准确性
5. 修复WebSocket协议结构体定义不匹配问题
6. 消除部分未使用变量编译警告
7. 修复openssl弃用警告的编译开关

+++ agentos/commons/utils/logging/src/service_logging.c
@@ -122,7 +122,7 @@ static bool level_filter_filter(filter_t* self, const log_record_t* record) {
     if (!self || !record) return false;
     int min_level = (int)(intptr_t)self->user_data;
     if (min_level <= 0) min_level = LOG_LEVEL_INFO;
-    return record->level >= min_level;
+    return (int)record->level >= min_level;
 }
 
 static void level_filter_destroy(filter_t* self) {
